use Methods enum to help TS fix error
to dynamicly evoke Router methods, for example
router.get(), router.put(). rather than hard code
the function name.
I add metadata(path and method) to handlers, but
when we retrieve the metadata back using Reflect.
getMetadata('method', target, key), the return
valur is any. so we have to define a enum and
constrain the type of the method to that enum.
because TS know all the Router methods, as long as
the enum is set within the range, there will be
no error.