include use decorator which will include middleware
use decorator is the same as route decorator,
it adds middleware as a metadate into handlers
when the controller decorator which has access to
the singleton retrieve metadate, up to now, path,
medthod, pathprefix, and middleware will be
retrieved. and added to the Router.
The index.ts will only use the Router.